server/pyproject.toml

[project]
#name = "safers-api"
#version = "0.1.0"
authors = [
    {name = "Astrosat", email = "info@astrosat.net"},
]
dependencies = [
    "argon2-cffi>=21.3.0",
    "boto3>=1.26.112",
    "django>=4.2",
    "django-colorfield>=0.8.0",
    "django-cors-headers>=3.14.0",
    "django-dbbackup>=4.0.2",
    "django-environ>=0.10.0",
    "django-filter>=23.1",
    "django-model-utils>=4.3.1",
    "django-on-heroku>=1.1.2",
    "django-sequences>=2.7",
    "django-silk>=5.0.3",
    "django-storages>=1.13.2",
    "dj-database-url>=1.3.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-gis>=1.0",
    "drf-spectacular[sidecar]>=0.26.1",
    "fusionauth-client>=1.45.0",
    "geopandas>=0.12.2",
    "jsonschema>=4.17.3",
    "pandas>=2.0.0",
    "pika>=1.3.1",
    "pillow>=9.5.0",
    "psycopg2-binary>=2.9.6",
    "tzdata>=2023.3",
    # not-needed
    "django-anymail[sendgrid,sparkpost]>=10.0",
    "drf-yasg>=1.21.5",
    "gunicorn>=20.1.0",
    "kombu>=5.2.4",
    "sentry-sdk>=1.19.1",
    "snakeviz>=2.2.0",
]
requires-python = ">=3.10"

[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "pylint",
    "yapf",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinxcontrib-confluencebuilder",
    "sphinxcontrib-django",
    "myst-parser",
    "python-dotenv",
]
test = [
    "pytest",
    "pytest-django",
    "pytest-html-reporter",
    "pytest-sugar",
    "factory-boy",
]

[tool.pdm.scripts]
docs = {shell = "cd ../docs && make html"}
test = "pytest"
